Create For_Loop_in_C
For each integer n in the interval [a,b] (given as input) :

If 1 <= n <= 9, then print the English representation of it in lowercase. That is "one" for , "two" for , and so on.
Else if n>9 and it is an even number, then print "even".
Else if n>9  and it is an odd number, then print "odd".