Create Calculate_The_Nth_Term
There is a series,S , where the next term is the sum of pervious three terms. Given the first three terms of the series,a ,b , and c respectively, you have to output the nth term of the series using recursion.

Recursive method for calculating nth term is given below.